LayoutTests/platform/mac-catalina/fast/text/international/system-language/navigator-language/navigator-language-hi-expected.txt

FAIL navigator.language should be hi-IN. Was hi.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
